% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkPlots.R
\name{chunk_labels}
\alias{chunk_labels}
\title{Generate labels for chunk facet plots}
\usage{
chunk_labels(dsChunk, gas_column, y_multiplier)
}
\arguments{
\item{dsChunk}{A dataframe containing all measurement data grouped by chunks for all gases.}

\item{gas_column}{Numeric. The column in the dsChunk with gas measurement data (e.g., 'CO2_dry').}

\item{y_multiplier}{Numeric. This multiplier is used for label placement.}
}
\value{
A list of labels for each chunk.
}
\description{
Generate labels for chunk facet plots
}
\examples{
labels_CO2 <- chunk_labels(dsChunk, CO2_dry, 1.05)
}
