% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkPlots.R
\name{chunk_plot}
\alias{chunk_plot}
\title{Generate facet plot with all chunks for selected gases. Please first create the labels for each gas using 'chunk_labels()'.}
\usage{
chunk_plot(dsChunk, gas_column, labels)
}
\arguments{
\item{dsChunk}{A dataframe containing all measurement data grouped by chunks for all gases.}

\item{gas_column}{Numeric. The column in the dsChunk with gas measurement data (e.g., 'CO2_dry').}

\item{labels}{A list of labels for each chunk.}
}
\value{
A ggplot object.
}
\description{
Generate facet plot with all chunks for selected gases. Please first create the labels for each gas using 'chunk_labels()'.
}
\examples{
Chunk_Facet_Plot_CO2 <- chunk_plot(dsChunk, CO2_dry, labels_CO2)
}
