% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToRegularTimesteps.R
\name{convertToRegularTimesteps}
\alias{convertToRegularTimesteps}
\title{Interpolate measurement data to equidistant timesteps}
\usage{
convertToRegularTimesteps(df, var_names, regular_timesteps)
}
\arguments{
\item{df}{A dataframe containing the measurement data and time stamps}

\item{var_names}{Character. The names of the variables that need to be interpolated to regular time steps}

\item{regular_timesteps}{A continuous timestamp vector}
}
\value{
A dataframe with equidistant timesteps
}
\description{
Interpolate measurement data from irregular timesteps to equidistant timesteps
}
\examples{
ds <- dataset
regular_timesteps <- seq(min(ds$TIMESTAMP),max(ds$TIMESTAMP), by="1 sec")
ds <-convertToRegularTimesteps(ds,c("CO2_dry", "H2O"),regular_timesteps)
}
