% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdditionalWeatherVariables.R
\name{getAdditionalWeatherVariables}
\alias{getAdditionalWeatherVariables}
\title{Get Additional Environmental Variables}
\usage{
getAdditionalWeatherVariables(latDeciDeg, lonDeciDeg, starttime, endtime)
}
\arguments{
\item{latDeciDeg}{Numeric. Latitude in decimal degrees, e.g., 37.915.}

\item{lonDeciDeg}{Numeric. Longitude in decimal degrees, e.g., -4.72.}

\item{starttime}{Character. Start date in '\%Y-\%m-\%d' format, e.g., '2022-05-01'.}

\item{endtime}{Character. End date in '\%Y-\%m-\%d' format, e.g., '2022-05-02'.}
}
\value{
A tibble containing the following environmental variables:
\itemize{
\item \code{h}: Elevation at the sample location. (m a.s.l)
\item \code{AirTemp}: Air temperature at 2m above ground. (°C)
\item \code{Pa}: Atmospheric pressure at the sample location elevation. (Pa)
\item \code{rel_humidity}: Relative humidity at 2m above ground. (\%)
\item \code{shortwave_radiation}: Shortwave radiation (W/m²)
\item \code{ET0}: Reference Evapotranspiration of a well-watered grass field (mm)
}
}
\description{
This function retrieves additional environmental variables for specific chamber setups where certain data might be missing.
The function uses the packages 'openmeteo' for atmospheric  and 'elevatr' for elevation data. Note that weather variables (hourly intervals) can only be retrieved for dates more than 5 days old.
}
\examples{
Additional_Weather_Data <- getAdditionalWeatherVariables(latDeciDeg = 37.915, lonDeciDeg = -4.72, starttime = '2022-05-01', endtime = '2022-05-02')

}
\references{
\itemize{
\item Zippenfenig, P. (2023). Open-Meteo.com Weather API \link{Computer software}. Zenodo. https://doi.org/10.5281/ZENODO.7970649
\item Hersbach, H., Bell, B., Berrisford, P., Biavati, G., Horányi, A., Muñoz Sabater, J., Nicolas, J., Peubey, C., Radu, R., Rozum, I., Schepers, D., Simmons, A., Soci, C., Dee, D., & Thépaut, J.-N. (2023). ERA5 hourly data on single levels from 1940 to present \link{Data set}. ECMWF. https://doi.org/10.24381/cds.adbb2d47
\item Muñoz Sabater, J. (2019). ERA5-Land hourly data from 2001 to present \link{Data set}. ECMWF. https://doi.org/10.24381/CDS.E2161BAC
\item Schimanke, S., Ridal, M., Le Moigne, P., Berggren, L., Undén, P., Randriamampianina, R., Andrea, U., Bazile, E., Bertelsen, A., Brousseau, P., Dahlgren, P., Edvinsson, L., El Said, A., Glinton, M., Hopsch, S., Isaksson, L., Mladek, R., Olsson, E., Verrelle, A., & Wang, Z.Q. (2021). CERRA sub-daily regional reanalysis data for Europe on single levels from 1984 to present \link{Data set}. ECMWF. https://doi.org/10.24381/CDS.622A565A
}
}
