% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDurationUncertaintyRelSD.R
\name{plotDurationUncertaintyRelSD}
\alias{plotDurationUncertaintyRelSD}
\title{plot the increase of uncertainty (Coefficient of variation) with decreasing measurement duration}
\usage{
plotDurationUncertaintyRelSD(
  ds,
  colTime = "TIMESTAMP",
  fRegress = c(exp = regressFluxExp, lin = regressFluxLinear, tanh = regressFluxTanh),
  ...,
  durations = seq(max(65, resFit0$tLag), max(times0), length.out = nDur + 1),
  nDur = 5,
  maxSdFluxRel = 0.5
)
}
\arguments{
\item{ds}{a data.frame}

\item{colTime}{column name of time \link{s}}

\item{fRegress}{the regression type}

\item{...}{further arguments to \code{\link{calcClosedChamberFlux}}}

\item{durations}{a vector with the durations to test}

\item{nDur}{number of durations to check}

\item{maxSdFluxRel}{maximum allowed relative standard deviation (=sd flux/ median flux)}
}
\value{
A Tibble
}
\description{
plot the increase of uncertainty (Coefficient of variation) with decreasing measurement duration
}
\examples{
resDur <- plotDurationUncertaintyRelSD( df, colConc = "CO2_dry", colTemp="AirTemp", volume = chamberVol,
fRegress = c(lin = regressFluxLinear, tanh = regressFluxTanh)
, maxSdFluxRel = 1
, durations = seq(60,max(as.numeric(df$TIMESTAMP) - as.numeric(df$TIMESTAMP[1])),20)
)
}
