% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStabilizationPoint.R
\name{findStabilizationPoint}
\alias{findStabilizationPoint}
\title{Find the stabilization Point of a curve}
\usage{
findStabilizationPoint(x, y)
}
\arguments{
\item{x}{A vector with durations}

\item{y}{A vector with the Fluxes, coefficients of variations or standard deviations}
}
\value{
A tibble with the Duration and Flux or SD or CV value where the curve stabilizes
}
\description{
Find the stabilization Point of a curve
}
\examples{
y <- resDur$statAll[[1]]$sdFlux / resDur$statAll[[1]]$fluxMedian
x <- resDur$statAll[[1]]$duration
result <- find_stabilization_point(x, y)

}
